Create controller on backend for data transfer

+++ or_atp/src/main/java/com/mr/or_atp/or_atp/models/automatic_train_protection/Modes/ATP_ModeInterface.java
@@ -9,9 +9,9 @@ public interface ATP_ModeInterface {
     void run(SimData simData);
     boolean isTractionAllowed();
     boolean isEmergencyBrakeForced();
-    int getObjectiveSpeed();
-    int getAllowedSpeed();
-    int getCurrentSpeed();
+    double getObjectiveSpeed();
+    double getAllowedSpeed();
+    double getCurrentSpeed();
     ATP_Panel getPanelIndications();
